Create base file for BaseJWTSyncAuth and BaseJJWTAsyncAuth
